ref(tiller): breakout ReleaseServer APIs into logical files.
*ReleaseServer.RunReleaseTest -> tiller/release_testing.go
*ReleaseServer.GetVersion     -> tiller/release_version.go
